Peel off client/hosted settings into its own context
