+++ apps/data/lib/grapevine_data/game_settings/hosted_settings.ex
@@ -1,4 +1,4 @@
-defmodule GrapevineData.Games.HostedSettings do
+defmodule GrapevineData.GameSettings.HostedSettings do
   @moduledoc """
   Client settings Schema
   """
