+++ apps/data/lib/grapevine_data/games.ex
@@ -6,10 +6,8 @@ defmodule GrapevineData.Games do
   alias GrapevineData.Accounts.User
   alias GrapevineData.Alerts
   alias GrapevineData.Filter
-  alias GrapevineData.Games.ClientSettings
   alias GrapevineData.Games.Connection
   alias GrapevineData.Games.Game
-  alias GrapevineData.Games.HostedSettings
   alias GrapevineData.Games.Images
   alias GrapevineData.Games.RedirectURI
   alias GrapevineData.Notifications
@@ -610,80 +608,6 @@ defmodule GrapevineData.Games do
 
       {:error, changeset} ->
         {:error, changeset}
-    end
-  end
-
-  @doc """
-  Edit the client settings
-  """
-  def edit_client_settings(game) do
-    game = Repo.preload(game, [:client_settings])
-
-    case is_nil(game.client_settings) do
-      true ->
-        game
-        |> Ecto.build_assoc(:client_settings)
-        |> ClientSettings.changeset(%{})
-
-      false ->
-        ClientSettings.changeset(game.client_settings, %{})
-    end
-  end
-
-  @doc """
-  Update web client settings for the game
-  """
-  def update_client_settings(game, params) do
-    game = Repo.preload(game, [:client_settings])
-
-    case is_nil(game.client_settings) do
-      true ->
-        game
-        |> Ecto.build_assoc(:client_settings)
-        |> ClientSettings.changeset(params)
-        |> Repo.insert()
-
-      false ->
-        game.client_settings
-        |> ClientSettings.changeset(params)
-        |> Repo.update()
-    end
-  end
-
-  @doc """
-  Edit the hosted settings
-  """
-  def edit_hosted_settings(game) do
-    game = Repo.preload(game, [:hosted_settings])
-
-    case is_nil(game.hosted_settings) do
-      true ->
-        game
-        |> Ecto.build_assoc(:hosted_settings)
-        |> HostedSettings.changeset(%{})
-
-      false ->
-        HostedSettings.changeset(game.hosted_settings, %{})
-    end
-  end
-
-  @doc """
-  Update web hosted settings for the game
-  """
-  def update_hosted_settings(game, params) do
-    game = Repo.preload(game, [:hosted_settings])
-
-    case is_nil(game.hosted_settings) do
-      true ->
-        game
-        |> Ecto.build_assoc(:hosted_settings)
-        |> HostedSettings.changeset(params)
-        |> Repo.insert()
-
-      false ->
-        game.hosted_settings
-        |> HostedSettings.changeset(params)
-        |> Repo.update()
     end
   end
