+++ apps/data/lib/grapevine_data/games/game.ex
@@ -11,10 +11,10 @@ defmodule GrapevineData.Games.Game do
   alias GrapevineData.Achievements.Achievement
   alias GrapevineData.Events.Event
   alias GrapevineData.Games
-  alias GrapevineData.Games.ClientSettings
   alias GrapevineData.Games.Connection
-  alias GrapevineData.Games.HostedSettings
   alias GrapevineData.Games.RedirectURI
+  alias GrapevineData.GameSettings.ClientSettings
+  alias GrapevineData.GameSettings.HostedSettings
   alias GrapevineData.Gauges.Gauge
   alias GrapevineData.Statistics.Session
