+++ apps/data/test/grapevine_data/games_test.exs
@@ -362,38 +362,5 @@ defmodule GrapevineData.GamesTest do
     end
   end
 
-  describe "update settings" do
-    test "when none exist" do
-      game = create_game(create_user())
-
-      {:ok, client_settings} = Games.update_client_settings(game, %{
-        character_package: "Char 0",
-        character_message: "Char.Status",
-        character_name_path: "name"
-      })
-
-      assert client_settings.game_id == game.id
-      assert client_settings.character_package == "Char 0"
-      assert client_settings.character_message == "Char.Status"
-      assert client_settings.character_name_path == "name"
-    end
-
-    test "updating existing" do
-      game = create_game(create_user())
-
-      {:ok, _client_settings} = Games.update_client_settings(game, %{
-        character_package: "Char 0",
-        character_message: "Char.Status",
-        character_name_path: "name"
-      })
-
-      {:ok, client_settings} = Games.update_client_settings(game, %{
-        character_name_path: "full_name"
-      })
-
-      assert client_settings.character_name_path == "full_name"
-    end
-  end
-
   defp connection_hook(_connection), do: :ok
 end
