+++ lib/web/controllers/manage/client_controller.ex
@@ -2,6 +2,7 @@ defmodule Web.Manage.ClientController do
   use Web, :controller
 
   alias GrapevineData.Games
+  alias GrapevineData.GameSettings
   alias GrapevineData.Gauges
 
   def show(conn, %{"game_id" => id}) do
@@ -11,7 +12,7 @@ defmodule Web.Manage.ClientController do
       conn
       |> assign(:game, game)
       |> assign(:gauges, Gauges.for(game))
-      |> assign(:changeset, Games.edit_client_settings(game))
+      |> assign(:changeset, GameSettings.edit_client_settings(game))
       |> render("show.html")
     end
   end
@@ -20,7 +21,7 @@ defmodule Web.Manage.ClientController do
     %{current_user: user} = conn.assigns
     {:ok, game} = Games.get(user, id)
 
-    with {:ok, _client_settings} <- Games.update_client_settings(game, params) do
+    with {:ok, _client_settings} <- GameSettings.update_client_settings(game, params) do
       conn
       |> put_flash(:info, "Updated!")
       |> redirect(to: manage_game_client_path(conn, :show, game.id))
