+++ lib/web/controllers/manage/hosted_site_controller.ex
@@ -2,6 +2,7 @@ defmodule Web.Manage.HostedSiteController do
   use Web, :controller
 
   alias GrapevineData.Games
+  alias GrapevineData.GameSettings
 
   def show(conn, %{"game_id" => id}) do
     %{current_user: user} = conn.assigns
@@ -9,7 +10,7 @@ defmodule Web.Manage.HostedSiteController do
     with {:ok, game} <- Games.get(user, id) do
       conn
       |> assign(:game, game)
-      |> assign(:changeset, Games.edit_hosted_settings(game))
+      |> assign(:changeset, GameSettings.edit_hosted_settings(game))
       |> render("show.html")
     end
   end
@@ -18,7 +19,7 @@ defmodule Web.Manage.HostedSiteController do
     %{current_user: user} = conn.assigns
     {:ok, game} = Games.get(user, id)
 
-    with {:ok, _hosted_settings} <- Games.update_hosted_settings(game, params) do
+    with {:ok, _hosted_settings} <- GameSettings.update_hosted_settings(game, params) do
       conn
       |> put_flash(:info, "Updated!")
       |> redirect(to: manage_game_hosted_site_path(conn, :show, game.id))
